backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "cleanarr"
version = "0.1.0"
description = "Jellyfin deletion webhook that cascades cleanup to Arr apps and download clients."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.6,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic-settings>=2.5.2,<3.0.0",
  "uvicorn[standard]>=0.30.6,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.14.1,<2.0.0",
  "pytest>=8.3.4,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "respx>=0.22.0,<1.0.0",
  "ruff>=0.8.6,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cleanarr"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "."]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["cleanarr"]
